+++ SIAG.Domain/Armazenagem/Cadastro/Interfaces/IAreaArmazenagemRepository.cs
@@ -3,9 +3,9 @@ using SIAG.Domain.Armazenagem.Cadastro.Models;
 
 namespace SIAG.Domain.Armazenagem.Cadastro.Interfaces
 {
-    public interface IAreaArmazenagemRepository : IBaseRepository<AreaArmazenagem, int>
+    public interface IAreaArmazenagemRepository : IBaseRepository<AreaArmazenagem, long>
     {
         public Task<DadosPaginadosDTO<AreaArmazenagem>> GetListAsync(FiltroPaginacaoDTO dto);
-        public Task<List<SelectDTO<int>>> GetSelectAsync(FiltroPaginacaoDTO dto);
+        public Task<List<SelectDTO<long>>> GetSelectAsync(FiltroPaginacaoDTO dto);
     }
 }
